Исправление замечаний

+++ internal/repository/urlshorterrepository/url_shorter_repository.go
@@ -145,5 +145,9 @@ func (r *urlShorterRepository) GetUserURLs(ctx context.Context, userID string) (
 
 // DeleteBatch удаляет несколько URL пакетно.
 func (r *urlShorterRepository) DeleteBatch(ctx context.Context, shortURLs []string, userID string) error {
+	if len(shortURLs) == 0 {
+		return nil
+	}
+
 	return r.storage.DeleteBatch(ctx, shortURLs, userID)
 }
